# src/generated/java/com/portfolio/renter/api/UserApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (4.2.3).
 * https://openapi-generator.tech Do not edit the class manually.
 */
package com.portfolio.renter.api;

import com.portfolio.renter.api.model.User;
import io.swagger.annotations.*;
import java.util.List;
import javax.validation.Valid;
import javax.validation.constraints.*;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.bind.annotation.RequestParam;

@Validated
@Api(value = "User", description = "the User API")
public interface UserApi {

  /**
   * POST /user : Create user This can only be done by the logged in user.
   *
   * @param body Created user object (required)
   * @return successful operation (status code 200)
   */
  @ApiOperation(
      value = "Create user",
      nickname = "createUser",
      notes = "This can only be done by the logged in user.",
      tags = {
        "user",
      })
  @ApiResponses(value = {@ApiResponse(code = 200, message = "successful operation")})
  @RequestMapping(value = "/user", method = RequestMethod.POST)
  ResponseEntity<Void> createUser(
      @ApiParam(value = "Created user object", required = true) @Valid @RequestBody User body);

  /**
   * POST /user/createWithArray : Creates list of users with given input array
   *
   * @param body List of user object (required)
   * @return successful operation (status code 200)
   */
  @ApiOperation(
      value = "Creates list of users with given input array",
      nickname = "createUsersWithArrayInput",
      notes = "",
      tags = {
        "user",
      })
  @ApiResponses(value = {@ApiResponse(code = 200, message = "successful operation")})
  @RequestMapping(value = "/user/createWithArray", method = RequestMethod.POST)
  ResponseEntity<Void> createUsersWithArrayInput(
      @ApiParam(value = "List of user object", required = true) @Valid @RequestBody
          List<User> body);

  /**
   * POST /user/createWithList : Creates list of users with given input array
   *
   * @param body List of user object (required)
   * @return successful operation (status code 200)
   */
  @ApiOperation(
      value = "Creates list of users with given input array",
      nickname = "createUsersWithListInput",
      notes = "",
      tags = {
        "user",
      })
  @ApiResponses(value = {@ApiResponse(code = 200, message = "successful operation")})
  @RequestMapping(value = "/user/createWithList", method = RequestMethod.POST)
  ResponseEntity<Void> createUsersWithListInput(
      @ApiParam(value = "List of user object", required = true) @Valid @RequestBody
          List<User> body);

  /**
   * DELETE /user/{username} : Delete user This can only be done by the logged in user.
   *
   * @param username The name that needs to be deleted (required)
   * @return Invalid username supplied (status code 400) or User not found (status code 404)
   */
  @ApiOperation(
      value = "Delete user",
      nickname = "deleteUser",
      notes = "This can only be done by the logged in user.",
      tags = {
        "user",
      })
  @ApiResponses(
      value = {
        @ApiResponse(code = 400, message = "Invalid username supplied"),
        @ApiResponse(code = 404, message = "User not found")
      })
  @RequestMapping(value = "/user/{username}", method = RequestMethod.DELETE)
  ResponseEntity<Void> deleteUser(
      @ApiParam(value = "The name that needs to be deleted", required = true)
          @PathVariable("username")
          String username);

  /**
   * GET /user/{username} : Get user by user name
   *
   * @param username The name that needs to be fetched. Use user1 for testing. (required)
   * @return successful operation (status code 200) or Invalid username supplied (status code 400)
   *     or User not found (status code 404)
   */
  @ApiOperation(
      value = "Get user by user name",
      nickname = "getUserByName",
      notes = "",
      response = User.class,
      tags = {
        "user",
      })
  @ApiResponses(
      value = {
        @ApiResponse(code = 200, message = "successful operation", response = User.class),
        @ApiResponse(code = 400, message = "Invalid username supplied"),
        @ApiResponse(code = 404, message = "User not found")
      })
  @RequestMapping(
      value = "/user/{username}",
      produces = {"application/xml", "application/json"},
      method = RequestMethod.GET)
  ResponseEntity<User> getUserByName(
      @ApiParam(
              value = "The name that needs to be fetched. Use user1 for testing. ",
              required = true)
          @PathVariable("username")
          String username);

  /**
   * GET /user/login : Logs user into the system
   *
   * @param username The user name for login (required)
   * @param password The password for login in clear text (required)
   * @return successful operation (status code 200) or Invalid username/password supplied (status
   *     code 400)
   */
  @ApiOperation(
      value = "Logs user into the system",
      nickname = "loginUser",
      notes = "",
      response = String.class,
      tags = {
        "user",
      })
  @ApiResponses(
      value = {
        @ApiResponse(code = 200, message = "successful operation", response = String.class),
        @ApiResponse(code = 400, message = "Invalid username/password supplied")
      })
  @RequestMapping(
      value = "/user/login",
      produces = {"application/xml", "application/json"},
      method = RequestMethod.GET)
  ResponseEntity<String> loginUser(
      @NotNull
          @ApiParam(value = "The user name for login", required = true)
          @Valid
          @RequestParam(value = "username", required = true)
          String username,
      @NotNull
          @ApiParam(value = "The password for login in clear text", required = true)
          @Valid
          @RequestParam(value = "password", required = true)
          String password);

  /**
   * GET /user/logout : Logs out current logged in user session
   *
   * @return successful operation (status code 200)
   */
  @ApiOperation(
      value = "Logs out current logged in user session",
      nickname = "logoutUser",
      notes = "",
      tags = {
        "user",
      })
  @ApiResponses(value = {@ApiResponse(code = 200, message = "successful operation")})
  @RequestMapping(value = "/user/logout", method = RequestMethod.GET)
  ResponseEntity<Void> logoutUser();

  /**
   * PUT /user/{username} : Updated user This can only be done by the logged in user.
   *
   * @param username name that need to be updated (required)
   * @param body Updated user object (required)
   * @return Invalid user supplied (status code 400) or User not found (status code 404)
   */
  @ApiOperation(
      value = "Updated user",
      nickname = "updateUser",
      notes = "This can only be done by the logged in user.",
      tags = {
        "user",
      })
  @ApiResponses(
      value = {
        @ApiResponse(code = 400, message = "Invalid user supplied"),
        @ApiResponse(code = 404, message = "User not found")
      })
  @RequestMapping(value = "/user/{username}", method = RequestMethod.PUT)
  ResponseEntity<Void> updateUser(
      @ApiParam(value = "name that need to be updated", required = true) @PathVariable("username")
          String username,
      @ApiParam(value = "Updated user object", required = true) @Valid @RequestBody User body);
}